# api/assets/swagger.json
{
  "consumes": ["application/json"],
  "produces": ["application/json"],
  "schemes": ["http"],
  "swagger": "2.0",
  "info": {
    "description": "GK Circle PCS examination API",
    "version": "0.0.1-alpha"
  },
  "host": "127.0.0.1",
  "basePath": "/api",
  "paths": {
...
    "/v1/shared_quizzes": {
      "get": {
        "consumes": ["application/json"],
        "schemes": ["http", "https"],
        "tags": ["ShareQuiz"],
        "summary": "List shared quiz for perticular user (only shared with the user or shared by the user).",
        "operationId": "RequestListSharedQuizzes",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "Type",
            "name": "type",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/ResponseListSharedQuizzes"
          },
          "400": {
            "$ref": "#/responses/GenericResFailNotFound"
          },
          "500": {
            "$ref": "#/responses/GenericResError"
          }
        }
      },
      "post": {
        "consumes": ["application/json"],
        "schemes": ["http", "https"],
        "tags": ["ShareQuiz"],
        "summary": "share quiz to other user.",
        "operationId": "RequestShareQuiz",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "QuizId",
            "name": "quiz_id",
            "in": "path",
            "required": true
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "x-go-name": "Email"
                },
                "permission": {
                  "type": "string",
                  "x-go-name": "Permission"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/ResponseOkWithMessage"
          },
          "400": {
            "$ref": "#/responses/GenericResFailNotFound"
          },
          "500": {
            "$ref": "#/responses/GenericResError"
          }
        }
      }
    },
    "/v1/shared_quizzes/{quiz_id}": {
      "get": {
        "consumes": ["application/json"],
        "schemes": ["http", "https"],
        "tags": ["ShareQuiz"],
        "summary": "List authorized users for perticular quiz.",
        "operationId": "RequestListQuizAuthorizedUsers",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "QuizId",
            "name": "quiz_id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/ResponseListQuizAuthorizedUsers"
          },
          "400": {
            "$ref": "#/responses/GenericResFailNotFound"
          },
          "500": {
            "$ref": "#/responses/GenericResError"
          }
        }
      },
      "put": {
        "consumes": ["application/json"],
        "schemes": ["http", "https"],
        "tags": ["ShareQuiz"],
        "summary": "Update authorized user permission for perticular quiz.",
        "operationId": "RequestUpdateUserPermissionOfQuiz",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "SharedQuizId",
            "name": "shared_quiz_id",
            "in": "query",
            "required": true
          },
          {
            "type": "string",
            "x-go-name": "QuizId",
            "name": "quiz_id",
            "in": "path",
            "required": true
          },
          {
            "name": "Body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "x-go-name": "Email"
                },
                "permission": {
                  "type": "string",
                  "x-go-name": "Permission"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/ResponseOkWithMessage"
          },
          "400": {
            "$ref": "#/responses/GenericResFailNotFound"
          },
          "500": {
            "$ref": "#/responses/GenericResError"
          }
        }
      },
      "delete": {
        "consumes": ["application/json"],
        "schemes": ["http", "https"],
        "tags": ["ShareQuiz"],
        "summary": "Delete authorized user permission for perticular quiz.",
        "operationId": "RequestDeleteUserPermissionOfQuiz",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "SharedQuizId",
            "name": "shared_quiz_id",
            "in": "query",
            "required": true
          },
          {
            "type": "string",
            "x-go-name": "QuizId",
            "name": "quiz_id",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/ResponseOkWithMessage"
          },
          "400": {
            "$ref": "#/responses/GenericResFailNotFound"
          },
          "500": {
            "$ref": "#/responses/GenericResError"
          }
        }
      }
    },
    "/v1/user/who": {
      "get": {
        "consumes": ["application/json"],
        "schemes": ["http", "https"],
        "tags": ["User"],
        "summary": "Get Details of user.",
        "operationId": "GetUserMeta",
        "responses": {
          "200": {
            "$ref": "#/responses/ResponseUserDetails"
          },
          "401": {
            "$ref": "#/responses/GenericResFailConflict"
          },
          "500": {
            "$ref": "#/responses/GenericResError"
          }
        }
      }
    },
    "/v1/user/{username}": {
      "post": {
        "consumes": ["application/json"],
        "schemes": ["http", "https"],
        "tags": ["User"],
        "summary": "Create Guest user to play for quiz directly without login.",
        "operationId": "RequestCreateQuickUser",
        "parameters": [
          {
            "type": "string",
            "x-go-name": "Username",
            "name": "username",
            "in": "path",
            "required": true
          },
          {
            "type": "string",
            "x-go-name": "Avatar",
            "name": "avatar_name",
            "in": "query",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/ResponseUserDetails"
          },
          "400": {
            "$ref": "#/responses/GenericResFailNotFound"
          },
          "500": {
            "$ref": "#/responses/GenericResError"
          }
        }
      }
    },
...
  },
  "definitions": {
    "NullString": {
      "description": "var s NullString\nerr := db.QueryRow(\"SELECT name FROM foo WHERE id=?\", id).Scan(\u0026s)\n...\nif s.Valid {\nuse s.String\n} else {\nNULL value\n}",
      "type": "object",
      "title": "NullString represents a string that may be null.\nNullString implements the Scanner interface so\nit can be used as a scan destination:",
      "properties": {
        "String": {
          "type": "string"
        },
        "Valid": {
          "type": "boolean"
        }
      },
      "x-go-package": "database/sql"
    },
...
    "UUID": {
      "description": "A UUID is a 128 bit (16 byte) Universal Unique IDentifier as defined in RFC\n4122.",
      "type": "array",
      "items": {
        "type": "integer",
        "format": "uint8"
      },
      "x-go-package": "github.com/google/uuid"
    }
  },
  "responses": {
    "GenericResError": {
      "description": "Unexpected error occurred",
      "schema": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "x-go-name": "Data"
          },
          "message": {
            "type": "string",
            "x-go-name": "Message"
          },
          "status": {
            "type": "string",
            "enum": ["error"],
            "x-go-name": "Status"
          }
        }
      }
    },
...
    "GenericResFailConflict": {
      "description": "Server understand request but refuse to authorize it",
      "schema": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "x-go-name": "Data"
          },
          "status": {
            "type": "string",
            "enum": ["fail"],
            "x-go-name": "Status"
          }
        }
      }
    },
    "GenericResFailNotFound": {
      "description": "Fail due to resource not exists",
      "schema": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "x-go-name": "Data"
          },
          "status": {
            "type": "string",
            "enum": ["fail"],
            "x-go-name": "Status"
          }
        }
      }
    },
...
    "ResponseListQuizAuthorizedUsers": {
      "description": "",
      "schema": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "first_name": {
                  "$ref": "#/definitions/NullString"
                },
                "id": {
                  "type": "string",
                  "x-go-name": "Id"
                },
                "img_key": {
                  "$ref": "#/definitions/NullString"
                },
                "last_name": {
                  "$ref": "#/definitions/NullString"
                },
                "permission": {
                  "type": "string",
                  "x-go-name": "Permission"
                },
                "shared_to": {
                  "type": "string",
                  "x-go-name": "SharedTo"
                }
              }
            },
            "x-go-name": "Data"
          },
          "status": {
            "type": "string",
            "x-go-name": "Status"
          }
        }
      }
    },
    "ResponseListSharedQuizzes": {
      "description": "",
      "schema": {
        "type": "object",
        "properties": {
          "data": {
            "type": "array",
            "items": {
              "type": "object",
              "properties": {
                "created_at": {
                  "type": "string",
                  "format": "date-time",
                  "x-go-name": "CreatedAt"
                },
                "creator_id": {
                  "type": "string",
                  "x-go-name": "CreatorID"
                },
                "description": {
                  "$ref": "#/definitions/NullString"
                },
                "id": {
                  "$ref": "#/definitions/UUID"
                },
                "title": {
                  "type": "string",
                  "x-go-name": "Title"
                },
                "total_questions": {
                  "type": "integer",
                  "format": "int64",
                  "x-go-name": "TotalQuestions"
                },
                "updated_at": {
                  "type": "string",
                  "format": "date-time",
                  "x-go-name": "UpdatedAt"
                }
              }
            },
            "x-go-name": "Data"
          },
          "status": {
            "type": "string",
            "x-go-name": "Status"
          }
        }
      }
    },
...
    "ResponseOkWithMessage": {
      "description": "",
      "schema": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "x-go-name": "Message"
          },
          "status": {
            "type": "string",
            "x-go-name": "Status"
          }
        }
      }
    },
...
    "ResponseUserDetails": {
      "description": "",
      "schema": {
        "type": "object",
        "properties": {
          "data": {
            "type": "object",
            "properties": {
              "created_at": {
                "type": "string",
                "x-go-name": "CreatedAt"
              },
              "email": {
                "type": "string",
                "x-go-name": "Email"
              },
              "first_name": {
                "type": "string",
                "x-go-name": "FirstName"
              },
              "id": {
                "type": "string",
                "x-go-name": "ID"
              },
              "img_key": {
                "type": "string",
                "x-go-name": "ImageKey"
              },
              "kratos_id": {
                "$ref": "#/definitions/NullString"
              },
              "last_name": {
                "type": "string",
                "x-go-name": "LastName"
              },
              "roles": {
                "type": "string",
                "x-go-name": "Roles"
              },
              "updated_at": {
                "type": "string",
                "x-go-name": "UpdatedAt"
              },
              "username": {
                "type": "string",
                "x-go-name": "Username"
              }
            },
            "x-go-name": "Data"
          },
          "status": {
            "type": "string",
            "x-go-name": "Status"
          }
        }
      }
    },
...
  }
}